requirements.txt
# Feishu OpenAPI SDK
lark-oapi>=1.4.0

# Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1
python-dotenv>=1.0.0

# Task Queue
celery>=5.3.4
redis>=7.0.0

# HTTP Client
requests>=2.31.0
urllib3>=2.1.0

# Utilities
gitpython>=3.1.40
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
tqdm>=4.66.1

# Development
pytest>=7.4.3
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0
async-timeout>=4.0.3
